+++ lib/src/test/java/com/auth0/spring/security/api/authentication/AuthenticationJsonWebTokenTest.java
@@ -4,7 +4,6 @@ import com.auth0.jwt.JWT;
 import com.auth0.jwt.JWTVerifier;
 import com.auth0.jwt.algorithms.Algorithm;
 import com.auth0.jwt.interfaces.DecodedJWT;
-import com.auth0.spring.security.api.authentication.AuthenticationJsonWebToken;
 import org.hamcrest.Matchers;
 import org.hamcrest.collection.IsCollectionWithSize;
 import org.hamcrest.collection.IsEmptyCollection;
@@ -15,6 +14,7 @@ import org.junit.rules.ExpectedException;
 import org.springframework.security.core.GrantedAuthority;
 
 import java.util.ArrayList;
+import java.util.Collection;
 import java.util.Collections;
 import java.util.Map;
 
@@ -208,13 +208,47 @@ public class AuthenticationJsonWebTokenTest {
 
         AuthenticationJsonWebToken auth = new AuthenticationJsonWebToken(token, verifier);
         assertThat(auth, is(notNullValue()));
+
+        Collection<? extends GrantedAuthority> authorities = auth.getAuthorities();
+        assertThat(authorities, is(notNullValue()));
+        assertThat(authorities, is(IsCollectionWithSize.hasSize(4)));
+        assertThat(authorities, containsInAnyOrder(
+                hasProperty("authority", is("auth0")),
+                hasProperty("authority", is("auth10")),
+                hasProperty("authority", is("SCOPE_auth0")),
+                hasProperty("authority", is("SCOPE_auth10"))
+        ));
+    }
+
+    @Test
+    public void shouldGetEmptyAuthoritiesOnEmptyPermissionsClaim() throws Exception {
+        String token = JWT.create()
+                .withArrayClaim("permissions", new String[]{})
+                .sign(hmacAlgorithm);
+
+        AuthenticationJsonWebToken auth = new AuthenticationJsonWebToken(token, verifier);
+        assertThat(auth, is(notNullValue()));
+        assertThat(auth.getAuthorities(), is(notNullValue()));
+        assertThat(auth.getAuthorities(), is(IsEmptyCollection.empty()));
+    }
+
+    @Test
+    public void shouldGetPermissionsAsAuthorities() throws Exception {
+        String[] permissionsClaim = {"read:permission", "write:permission"};
+        String token = JWT.create()
+                .withArrayClaim("permissions", permissionsClaim)
+                .sign(hmacAlgorithm);
+
+        AuthenticationJsonWebToken auth = new AuthenticationJsonWebToken(token, verifier);
+        assertThat(auth, is(notNullValue()));
         assertThat(auth.getAuthorities(), is(notNullValue()));
         assertThat(auth.getAuthorities(), is(IsCollectionWithSize.hasSize(2)));
 
-        ArrayList<GrantedAuthority> authorities = new ArrayList<>(auth.getAuthorities());
-        assertThat(authorities.get(0), is(notNullValue()));
-        assertThat(authorities.get(0).getAuthority(), is("auth0"));
-        assertThat(authorities.get(1), is(notNullValue()));
-        assertThat(authorities.get(1).getAuthority(), is("auth10"));
+        Collection<? extends GrantedAuthority> authorities = auth.getAuthorities();
+        assertThat(authorities, IsCollectionWithSize.hasSize(2));
+        assertThat(authorities, containsInAnyOrder(
+                hasProperty("authority", is("PERMISSION_" + permissionsClaim[0])),
+                hasProperty("authority", is("PERMISSION_" + permissionsClaim[1]))
+        ));
     }
 }